Refactor Next.js configuration by removing legacy config file, updating to ES module format, and enhancing PostCSS setup with new plugins. Update package dependencies for improved functionality and security.
