+++ package.json
@@ -46,9 +46,13 @@
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
+    "css-loader": "^7.1.2",
     "eslint": "^8",
     "eslint-config-next": "14.0.3",
     "postcss": "^8",
+    "postcss-loader": "^8.1.1",
+    "postcss-preset-env": "^10.1.5",
+    "style-loader": "^4.0.0",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
   }
