+++ postcss.config.mjs
@@ -2,6 +2,8 @@
 const config = {
   plugins: {
     tailwindcss: {},
+    autoprefixer: {},
+    'postcss-preset-env': {},
   },
 };
